igvmbuilder: Refactor igvmbuilder into separate crates
The implementation suffered from being a direct converstion from the the
C builder. This commit separates the code into different structs/crates
and uses rust more effectively.
